computer-use-demo/PyAutoGUIControl/requirements.txt
pyautogui 
python-dotenv
pyperclip
